lichenhui/crowdfunding/requirements.txt
Django==1.9
django-pure-pagination==0.3.0
mysqlclient==1.3.12
Pillow==5.1.0
PyMySQL==0.8.1
